# openapi.json
{
  "components": {
    "schemas": {
      "ActiveRunUnavailableResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ActiveRunUnavailableResponse",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResetSettingsRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "const": "RESET TO PROTECTED DEFAULTS",
            "title": "Confirmation",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "ResetSettingsRequest",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": false,
        "properties": {
          "achieved_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Achieved Count",
            "type": "integer"
          },
          "admitted_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Admitted Count",
            "type": "integer"
          },
          "capacity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunCapacity"
              },
              {
                "type": "null"
              }
            ]
          },
          "committed_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Committed Count",
            "type": "integer"
          },
          "committed_opportunity_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 8,
            "title": "Committed Opportunity Ids",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_stage": {
            "default": "created",
            "title": "Current Stage",
            "type": "string"
          },
          "effective_config": {
            "additionalProperties": true,
            "title": "Effective Config",
            "type": "object"
          },
          "evaluated_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Evaluated Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "limit_code": {
            "anyOf": [
              {
                "enum": [
                  "queries",
                  "search_results",
                  "unique_urls",
                  "retrieved_pages",
                  "retrieved_bytes",
                  "source_bytes",
                  "signals",
                  "patterns",
                  "segments",
                  "preliminary_survivors",
                  "concepts",
                  "opportunities",
                  "search_retries",
                  "model_calls",
                  "repairs",
                  "total_tokens",
                  "role_tokens"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Code"
          },
          "limit_stage": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[a-z0-9_]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Stage"
          },
          "model_usage": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ModelCallUsage"
            },
            "maxItems": 256,
            "title": "Model Usage",
            "type": "array"
          },
          "non_counting_outcomes": {
            "$ref": "#/components/schemas/RunNonCountingOutcomes"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "shortfall_code": {
            "anyOf": [
              {
                "enum": [
                  "bounded_pool_exhausted",
                  "insufficient_evidence"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortfall Code"
          },
          "shortfall_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortfall Detail"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "target_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Target Count",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "warning_codes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warning Codes",
            "type": "array"
          },
          "work_counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Work Counters",
            "type": "object"
          },
          "workflow_version": {
            "title": "Workflow Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "opportunity_id",
          "effective_config",
          "workflow_version",
          "schema_version",
          "created_at"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "StartRunConflictResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/StartBlockedDetail"
              }
            ],
            "title": "Detail"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "detail"
        ],
        "title": "StartRunConflictResponse",
        "type": "object"
      },
      "StartRunRequest": {
        "additionalProperties": false,
        "properties": {},
        "title": "StartRunRequest",
        "type": "object"
      },
      "StartRunResponse": {
        "properties": {
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "state",
          "opportunity_id",
          "warnings"
        ],
        "title": "StartRunResponse",
        "type": "object"
      },
      "StopRunRequest": {
        "additionalProperties": false,
        "properties": {},
        "title": "StopRunRequest",
        "type": "object"
      },
      "StopRunResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "accepted"
        ],
        "title": "StopRunResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pensae Signal",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "post": {
        "operationId": "start_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRunConflictResponse"
                }
              }
            },
            "description": "Capability preflight blocked start or another run is active"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Research service unavailable"
          }
        },
        "summary": "Start Run"
      }
    },
    "/api/runs/active": {
      "get": {
        "operationId": "active_run_snapshot_api_runs_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No active research run"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveRunUnavailableResponse"
                }
              }
            },
            "description": "Research service or active snapshot unavailable"
          }
        },
        "summary": "Active Run Snapshot"
      }
    },
...
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "run_snapshot_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Snapshot"
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "operationId": "run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 48,
                  "pattern": "^[0-9]+-[0-9]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Events"
      }
    },
    "/api/runs/{run_id}/stop": {
      "post": {
        "operationId": "stop_run_api_runs__run_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Run"
      }
    },
    "/api/settings": {
      "get": {
        "operationId": "get_saved_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Saved Settings"
      },
      "put": {
        "operationId": "save_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Settings"
      }
    },
    "/api/settings/reset": {
      "post": {
        "operationId": "reset_settings_api_settings_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Settings"
      }
    },
...
  }
}